backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
clickhouse-driver
sqlalchemy
